Updated for serialization
overrided writeObject and ReadObject methods so that the mainFrame instance variable is properly set up, new helper method inScrollSetUp helps with this process. Made mainFrame transient because the ActionListener attached to upload kept attempting to serialize the JFileChooser for uploading, which is not serializable .
Added checks and filters so only image files can be uploaded.